+++ cmd/internal/docker_test.go
@@ -9,13 +9,13 @@ import (
 
 func TestEvaluateDockerComposeStatus(t *testing.T) {
 	// Mock the BloodHound Docker YAML files
-	homeDir := GetDefaultHomeDir()
-	err := os.MkdirAll(homeDir, 0755)
+	dataDir := GetDefaultDataDir()
+	err := os.MkdirAll(dataDir, 0755)
 	assert.NoError(t, err, "Expected directory creation to succeed")
 
-	localMockYaml := filepath.Join(GetDefaultHomeDir(), "docker-compose.dev.yml")
+	localMockYaml := filepath.Join(GetDefaultDataDir(), "docker-compose.dev.yml")
 	local, localErr := os.Create(localMockYaml)
-	prodMockYaml := filepath.Join(GetDefaultHomeDir(), "docker-compose.yml")
+	prodMockYaml := filepath.Join(GetDefaultDataDir(), "docker-compose.yml")
 	prod, prodErr := os.Create(prodMockYaml)
 	assert.Equal(t, nil, localErr, "Expected `os.Create()` to return no error")
 	assert.Equal(t, nil, prodErr, "Expected `os.Create()` to return no error")